cleanup(virtctl,tests): Move virtctl vnc tests
Move the virtctl vnc tests to tests/virtctl.
